requirements.txt
Django==1.9.4
argparse==1.2.1
pyenchant==1.6.6
wsgiref==0.1.2